V0.1.6.7 修复 bug

+++ env_init/src/uninstall_vault.sh
@@ -12,12 +12,14 @@ if [[ $1 = -n ]]; then noenter=true; else noenter=false; fi
 echo -e "${b}------------------- Vault 卸载 -------------------${n}"
 
 echo -e "${b}正在卸载 Vault...${n}"
-sudo systemctl stop vault
-expect << EOF
+if sudo systemctl is-active --quiet vault; then sudo systemctl stop vault; fi
+if dpkg -l | grep -q vault; then
+    expect << EOF
 spawn sudo apt-get remove --purge vault
 expect "Do you want to continue?" { send "Y\r" }
 expect eof
 EOF
+fi
 sudo rm -rf /etc/vault.d
 if [[ $noenter = true ]]; then ans=y
 else echo -e -n "${r}删除 Vault 的数据？(y/n): ${n}"; read -p "" ans; fi
@@ -25,9 +27,8 @@ if [[ $ans = y ]]; then
     echo -e "${b}删除所有 Vault 的数据.${n}"
     sudo rm -rf /opt/vault
 fi
-sudo systemctl daemon-reload
 
-rm -f /home/woo/.env
+sudo systemctl daemon-reload
 
 echo -e "${g}Vault 已成功卸载。${n}"
